major refactor, renaming to q2_matchmaker
