+++ setup.py
@@ -10,7 +10,7 @@ classes = """
     Programming Language :: Python :: 3
     Programming Language :: Python :: 3 :: Only
     Programming Language :: Python :: 3.4
-    Programming Language :: Python :: 3.5
+n    Programming Language :: Python :: 3.5
     Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Operating System :: Unix
@@ -19,7 +19,7 @@ classes = """
 """
 classifiers = [s.strip() for s in classes.split('\n') if s]
 
-description = ('QIIME2 plugin for Differential Abundance Analysis.')
+description = ('QIIME2 plugin for Case-Control Differential Abundance Analysis.')
 
 
 setup(name='q2-differential',
@@ -37,10 +37,10 @@ setup(name='q2-differential',
           'matplotlib',
       ],
       entry_points={
-          'qiime2.plugins': ['q2-differential=q2_differential.plugin_setup:plugin']
+          'qiime2.plugins': ['q2-matchmaker=q2_matchmaker.plugin_setup:plugin']
       },
       package_data={
-          "q2_differential": ['assets/nb_case_control.stan'],
+          "q2_matchmaker": ['assets/nb_case_control.stan'],
       },
       scripts=glob('scripts/*'),
       classifiers=classifiers,
